+++ src/core/scanlators/api_based.py
@@ -1,3 +1,4 @@
+import hashlib
 import re
 from types import SimpleNamespace
 from typing import Optional
@@ -47,11 +48,11 @@ class _Comick(AbstractScanlator):
 
     async def get_synopsis(self, raw_url: str) -> str:
         manga_id = await self.get_id(raw_url)
-        return await self.bot.comick_api.get_synopsis(manga_id)
+        return await self.bot.apis.comick.get_synopsis(manga_id)
 
     async def get_all_chapters(self, raw_url: str) -> list[Chapter] | None:
         manga_id = await self.get_id(raw_url)
-        chapters = await self.bot.comick_api.get_chapters_list(manga_id)
+        chapters = await self.bot.apis.comick.get_chapters_list(manga_id)
         if chapters:
             url_name = self.rx.search(raw_url).groupdict()["url_name"]
             return [
@@ -67,7 +68,7 @@ class _Comick(AbstractScanlator):
 
     async def get_title(self, raw_url: str) -> str | None:
         manga_id = await self.get_id(raw_url)
-        manga = await self.bot.comick_api.get_manga(manga_id)
+        manga = await self.bot.apis.comick.get_manga(manga_id)
         if manga.get("statusCode", 200) == 404:
             return None
         return manga["comic"]["title"]
@@ -80,7 +81,7 @@ class _Comick(AbstractScanlator):
 
     async def get_status(self, raw_url: str) -> str:
         manga_id = await self.get_id(raw_url)
-        manga = await self.bot.comick_api.get_manga(manga_id)
+        manga = await self.bot.apis.comick.get_manga(manga_id)
         if manga.get("statusCode", 200) == 404:
             return "Unknown"
         status_map = {1: "Ongoing", 2: "Completed", 3: "Cancelled", 4: "Hiatus"}
@@ -95,10 +96,10 @@ class _Comick(AbstractScanlator):
 
     async def get_cover(self, raw_url: str) -> str | None:
         manga_id = await self.get_id(raw_url)
-        return await self.bot.comick_api.get_cover(manga_id)
+        return await self.bot.apis.comick.get_cover(manga_id)
 
     async def search(self, query: str, as_em: bool = False) -> list[PartialManga] | list[discord.Embed]:
-        json_resp: list[dict] = await self.bot.comick_api.search(query=query)
+        json_resp: list[dict] = await self.bot.apis.comick.search(query=query)
         found_manga: list[PartialManga] = []
         if not json_resp:
             return found_manga
@@ -153,7 +154,7 @@ class _MangaDex(AbstractScanlator):
 
     async def get_synopsis(self, raw_url: str) -> str:
         manga_id = await self.get_id(raw_url)
-        synopsis = await self.bot.mangadex_api.get_synopsis(manga_id)
+        synopsis = await self.bot.apis.mangadex.get_synopsis(manga_id)
         if synopsis:
             return synopsis
         else:
@@ -161,7 +162,7 @@ class _MangaDex(AbstractScanlator):
 
     async def get_all_chapters(self, raw_url: str) -> list[Chapter] | None:
         manga_id = await self.get_id(raw_url)
-        chapters = await self.bot.mangadex_api.get_chapters_list(manga_id)
+        chapters = await self.bot.apis.mangadex.get_chapters_list(manga_id)
         if chapters:
             return [
                 Chapter(
@@ -174,13 +175,13 @@ class _MangaDex(AbstractScanlator):
 
     async def get_status(self, raw_url: str) -> str:
         manga_id = await self.get_id(raw_url)
-        manga = await self.bot.mangadex_api.get_manga(manga_id)
+        manga = await self.bot.apis.mangadex.get_manga(manga_id)
         status = manga["data"]["attributes"]["status"].lower().capitalize()
         return status
 
     async def get_title(self, raw_url: str) -> str | None:
         manga_id = await self.get_id(raw_url)
-        manga = await self.bot.mangadex_api.get_manga(manga_id)
+        manga = await self.bot.apis.mangadex.get_manga(manga_id)
         return manga["data"]["attributes"]["title"]["en"]
 
     async def get_id(self, raw_url: str) -> str:
@@ -196,15 +197,15 @@ class _MangaDex(AbstractScanlator):
 
     async def get_cover(self, raw_url: str) -> str | None:
         manga_id = await self.get_id(raw_url)
-        manga = await self.bot.mangadex_api.get_manga(manga_id)
+        manga = await self.bot.apis.mangadex.get_manga(manga_id)
         cover_id = [
             x["id"] for x in manga["data"]["relationships"] if x["type"] == "cover_art"
         ][0]
-        cover_url = await self.bot.mangadex_api.get_cover(manga["data"]["id"], cover_id)
+        cover_url = await self.bot.apis.mangadex.get_cover(manga["data"]["id"], cover_id)
         return cover_url
 
     async def search(self, query: str, as_em: bool = False) -> list[PartialManga] | list[discord.Embed]:
-        json_resp = await self.bot.mangadex_api.search(title=query)
+        json_resp = await self.bot.apis.mangadex.search(title=query)
         found_manga: list[PartialManga] = []
         if not json_resp["data"]:
             return found_manga
@@ -221,7 +222,148 @@ class _MangaDex(AbstractScanlator):
                 if rel_dict["type"] == "cover_art":
                     cover_id = rel_dict["id"]
                     break
-            cover = await self.bot.mangadex_api.get_cover(manga_id=_id, cover_id=cover_id)
+            cover = await self.bot.apis.mangadex.get_cover(manga_id=_id, cover_id=cover_id)
+            p_manga = PartialManga(_id, title, url, self.name, cover_url=cover)
+            found_manga.append(p_manga)
+        if as_em:
+            found_manga: list[discord.Embed] = self.partial_manga_to_embed(found_manga)
+        return found_manga
+
+
+class _ZeroScans(AbstractScanlator):
+    rx: re.Pattern = RegExpressions.zeroscans_url
+    icon_url = "https://zeroscans.com/favicon.ico"
+    base_url = "https://zeroscans.com"
+    fmt_url = base_url + "/comics/{url_name}"
+    chp_url_fmt = fmt_url + "/{chapter_id}"
+
+    def __init__(self, name: str):
+        super().__init__(name)
+        self.name: str = name
+
+    @property
+    def json_tree(self):
+        return SimpleNamespace(
+            properties=SimpleNamespace(
+                icon_url=self.icon_url,
+                base_url=self.base_url,
+                requires_update_embed=False,
+                can_render_cover=True,
+                format_urls=SimpleNamespace(
+                    manga=self.fmt_url,
+                )
+            )
+        )
+
+    async def get_fp_partial_manga(self) -> list[PartialManga]:
+        new_chapters: list[dict] = await self.bot.apis.zeroscans.get_latest_chapters()
+        found_manhwa: list[PartialManga] = []
+        for manhwa_dict in new_chapters:
+            url_name = manhwa_dict["slug"]
+            url = await self.format_manga_url(url_name=url_name)
+            manga_id = await self.get_id(url)
+            name = manhwa_dict["name"]
+            cover = manhwa_dict["cover"]["vertical"]
+            latest_chapters: list[Chapter] = []
+            for i, chapter_dict in enumerate(manhwa_dict["chapters"]):
+                chap = Chapter(
+                    self.chp_url_fmt.format(url_name=url_name, chapter_id=chapter_dict["id"]),
+                    f'Chapter {chapter_dict["name"]}',
+                    i
+                )
+                latest_chapters.append(chap)
+            p_manhwa = PartialManga(
+                manga_id, name, url, self.name, cover, latest_chapters
+            )
+            found_manhwa.append(p_manhwa)
+        return found_manhwa
+
+    def check_ownership(self, raw_url: str) -> bool:
+        return self.rx.search(raw_url) is not None
+
+    async def _get_url_name(self, raw_url: str) -> str:
+        try:
+            return self.rx.search(raw_url).groupdict().get("url_name")
+        except (AttributeError, TypeError) as e:
+            self.bot.logger.error(raw_url)
+            raise e
+
+    async def get_synopsis(self, raw_url: str) -> str:
+        url_name = await self._get_url_name(raw_url)
+        return await self.bot.apis.zeroscans.get_synopsis(url_name)
+
+    async def get_all_chapters(self, raw_url: str) -> list[Chapter] | None:
+        url_name = await self._get_url_name(raw_url)
+        manga_id = await self.bot.apis.zeroscans.get_manga_id(url_name)
+        chapters = await self.bot.apis.zeroscans.get_chapters_list(manga_id)
+        if chapters:
+            url_name = self.rx.search(raw_url).groupdict()["url_name"]
+            return [
+                Chapter(
+                    self.chp_url_fmt.format(
+                        url_name=url_name, chapter_id=chp["id"]), f'Chapter {chp["name"]}',
+                    i
+                )
+                for i, chp in enumerate(chapters)
+            ]
+        else:
+            return []
+
+    async def get_title(self, raw_url: str) -> str | None:
+        url_name = await self._get_url_name(raw_url)
+        manga = await self.bot.apis.zeroscans.get_manga(url_name)
+        return manga["data"]["name"]
+
+    async def get_id(self, raw_url: str) -> str | None:
+        try:
+            url_id = self.rx.search(raw_url).groupdict().get("id")
+        except (AttributeError, TypeError) as e:
+            self.bot.logger.error(raw_url)
+            raise e
+        if url_id is None or self.json_tree.properties.dynamic_url is True:
+            key = await self._get_url_name(raw_url)
+            return hashlib.sha256(key.encode()).hexdigest()
+        return url_id
+        # return await self.bot.apis.zeroscans.get_manga_id(await self._get_url_name(raw_url))
+
+    async def get_status(self, raw_url: str) -> str:
+        url_name = await self._get_url_name(raw_url)
+        manga = await self.bot.apis.zeroscans.get_manga(url_name)
+        for status_dict in manga["data"]["statuses"]:
+            if status_dict["slug"] == "ongoing":
+                return "Ongoing"
+        else:
+            return manga["data"]["statuses"][-1]["name"] if manga["data"]["statuses"] else "Completed"
+
+    async def format_manga_url(
+            self, raw_url: Optional[str] = None, url_name: Optional[str] = None, _: Optional[str] = None
+    ) -> str:
+        if not url_name:
+            url_name = self.rx.search(raw_url).groupdict()["url_name"]
+        return self.fmt_url.format(url_name=url_name)
+
+    async def get_cover(self, raw_url: str) -> str | None:
+        url_name = await self._get_url_name(raw_url)
+        return await self.bot.apis.zeroscans.get_cover(url_name)
+
+    async def search(self, query: str, as_em: bool = False) -> list[PartialManga] | list[discord.Embed]:
+        json_resp: list[dict] = await self.bot.apis.zeroscans.search(title=query, limit=10)
+        found_manga: list[PartialManga] = []
+        if not json_resp:
+            return found_manga
+
+        for item_dict in json_resp:
+            title = item_dict["name"]
+            _id = item_dict.get("id")
+            url_name = item_dict["slug"]
+            url = await self.format_manga_url(url_name=url_name)
+            for key in ["full", "vertical", "horizontal"]:
+                cover = item_dict["cover"].get(key)
+                if cover:
+                    break
+            else:
+                cover = None
+
             p_manga = PartialManga(_id, title, url, self.name, cover_url=cover)
             found_manga.append(p_manga)
         if as_em:
@@ -232,9 +374,11 @@ class _MangaDex(AbstractScanlator):
 class CustomKeys:
     comick: str = "comick"
     mangadex: str = "mangadex"
+    zeroscans: str = "zeroscans"
 
 
 keys = CustomKeys()
 
 scanlators[keys.comick] = _Comick(keys.comick)
 scanlators[keys.mangadex] = _MangaDex(keys.mangadex)
+scanlators[keys.zeroscans] = _ZeroScans(keys.zeroscans)