+++ src/core/scanlators/custom.py
@@ -34,6 +34,8 @@ class _ReaperScans(BasicScanlator):
             tuple[list[Chapter] | None, int] - the list of chapters and max number of chapters there are
         """
         req_url = manga_url.removesuffix("/")[:max(manga_url.find("?"), len(manga_url))]
+        if page is not None and page > 1:
+            req_url += f"?page={page}"
         text = await self._get_text(req_url, "GET")
         soup = BeautifulSoup(text, "html.parser")
         self.remove_unwanted_tags(soup, self.json_tree.selectors.unwanted_tags)
@@ -66,7 +68,6 @@ class _ReaperScans(BasicScanlator):
         return found_chapters, max_chapters
 
     async def get_all_chapters(self, raw_url: str, page: int | None = None) -> list[Chapter] | None:
-        #  FixMe: There's some bug that causes a chapter update to be sent 6 times?
         if page is not None and page == 0:
             return (await self.get_chapters_on_page(raw_url, page=page))[0]
 
@@ -99,46 +100,10 @@ class _OmegaScans(NoStatusBasicScanlator):
         super().__init__(name, **kwargs)
 
     async def search(self, query: str, as_em: bool = False) -> list[PartialManga] | list[discord.Embed]:
-        search_url = self.json_tree.search.url
-        if self.json_tree.search.query_parsing.encoding == "url":
-            query = url_encode(query)
-        elif self.json_tree.search.query_parsing.encoding is None:
-            for pattern_val_dict in self.json_tree.search.query_parsing.regex:
-                pattern = re.compile(pattern_val_dict["pattern"])
-                sub_value = pattern_val_dict["sub_value"]
-                query = pattern.sub(sub_value, query)
-        # if encoding is == "raw" then we do nothing
-        extra_params: dict = self.json_tree.search.extra_params
-        if self.json_tree.search.as_type == "path":
-            params = "?" + "&".join([f"{k}={v}" for k, v in extra_params.items() if v is not None])
-            null_params = {k: v for k, v in extra_params.items() if v is None}
-            if null_params:
-                params += "&".join([f"{k}" for k, v in null_params.items()])
-            query += params
-            request_kwargs = {
-                "url": search_url + query, "method": self.json_tree.search.request_method
-            }
-        else:  # as param
-            params = {self.json_tree.search.search_param_name: query} | extra_params
-            request_kwargs = {
-                "url": search_url, "params": params, "method": self.json_tree.search.request_method
-            }
-
-        if self.json_tree.request_method == "http":
-            # noinspection PyProtectedMember
-            resp = await self.bot.session._request(**request_kwargs)
-            await raise_and_report_for_status(self.bot, resp)
-            json_resp = await resp.json()
-        else:  # req method is "curl"
-            resp = await self.bot.curl_session.request(**request_kwargs)
-            await raise_and_report_for_status(self.bot, resp)
-            json_resp = resp.json()
-
-        # here we have the resp object
-
+        search_results = await self.bot.apis.omegascans.search(query)
         found_manga: list[PartialManga] = []
 
-        for item_dict in json_resp:
+        for item_dict in search_results:
             title = item_dict["title"]
             url = self.json_tree.properties.format_urls.manga.format(url_name=item_dict["series_slug"])
             cover = item_dict["thumbnail"]
@@ -149,6 +114,24 @@ class _OmegaScans(NoStatusBasicScanlator):
         if as_em:
             found_manga: list[discord.Embed] = self.partial_manga_to_embed(found_manga)
         return found_manga
+
+    async def get_all_chapters(self, raw_url: str) -> list[Chapter]:
+        url_name = await self._get_url_name(raw_url)
+        chapters: list[dict] = await self.bot.apis.omegascans.get_chapters_list(url_name)
+        found_chapters: list[Chapter] = []
+        for i, chapter in enumerate(chapters):
+            base_chapter_url = await self.format_manga_url(url_name=url_name)
+            chapter_url = base_chapter_url.removesuffix("/") + "/" + chapter["chapter_slug"]
+            chapter_name = chapter["chapter_name"]
+            if chapter["chapter_title"] is not None:
+                chapter_name += f" - {chapter['chapter_title']}"
+            found_chapters.append(Chapter(chapter_url, chapter_name, i))
+        return found_chapters
+
+    async def get_status(self, raw_url: str) -> str:
+        url_name = await self._get_url_name(raw_url)
+        status = await self.bot.apis.omegascans.get_status(url_name)
+        return status
 
 
 class _RealmScans(DynamicURLScanlator):